Remove need for passphrase w/ do-not-broadcast
Previously, when using the “do not broadcast” feature on transaction modals, you were still required to provide a passphrase.  The “raw transaction” dialog that apperars contains both signed and unsigned versions of the transaction bytes.  This commit allows you to use “do not broadcast” without providing the passphrase.  In such a case, the “raw transaciton” dialog will only show the unsigned transaction bytes.